requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
charset-normalizer==3.3.0
click==8.1.7
fastapi==0.104.0
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.1
idna==3.4
iso8601==1.1.0
loguru==0.7.2
multidict==6.0.4
pydantic==2.4.2
pydantic_core==2.10.1
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
yarl==1.9.2
